requirements.txt
Flask>=2.2.0
tweepy>=4.14.0
requests
python-dotenv
pandas
matplotlib
streamlit
plotly
wordcloud
scikit-learn
